fixup! lsp: Add console_scripts entry point
