+++ lib/esbonio/setup.cfg
@@ -38,7 +38,7 @@ exclude = tests*
 
 [options.entry_points]
 console_scripts =
-    esbonio = esbonio.__main__
+    esbonio = esbonio.cli:main
 
 [options.extras_require]
 dev = black ; flake8 ; pre-commit ; pytest ; pytest-cov ; tox
